requirements.txt
ghapi
emoji_data_python
yattag
pyyaml
